+++ FireworkVideoSample/FireworkVideoShoppingSample/View/AppTheme.swift
@@ -1,0 +1,15 @@
+//
+//  AppTheme.swift
+//  FireworkVideoShoppingSample
+//
+//  Copyright © 2021 Firework. All rights reserved.
+//
+
+import UIKit
+
+struct AppTheme {
+    static let backgroundColor = UIColor.systemBackground
+    static let labelColor = UIColor.label
+    static let ctaButtonColor = UIColor.systemRed
+}
+